requirements.txt
requests>=2.28.0
pdfplumber>=0.7.0
google-cloud-aiplatform>=1.25.0
pytz>=2022.1
python-dotenv>=1.0.0
beautifulsoup4>=4.11.1
lxml>=4.9.1
rarfile>=4.0 